Make update idempotent

+++ custom_updates/0.3.6_updates.sh
@@ -8,10 +8,10 @@ echo "Stopping hornet for uid/gid change ..."
 /bin/systemctl stop hornet
 
 echo "Changing uid/gid for hornet to match docker image ..."
-usermod -u 39999 hornet
-groupmod -g 39999 hornet
+usermod -u 39999 hornet || true
+groupmod -g 39999 hornet || true
 mkdir -p /var/lib/hornet/snapshot
-mv -- /var/lib/hornet/latest-export.gz.bin /var/lib/hornet/snapshot/.
+[ -f /var/lib/hornet/latest-export.gz.bin ] && mv -- /var/lib/hornet/latest-export.gz.bin /var/lib/hornet/snapshot/.
 chown hornet:hornet /var/lib/hornet -R
 
 echo "Set new path for latest-export.gz.bin ..."
